Decouple behaviour of CLI from formatting
