+++ packages/cli/src/commands/lint/format.ts
@@ -1,10 +1,5 @@
-import { LintResult } from "@oursky/gitlint";
 import chalk from "chalk";
-
-type LintReportResults = LintReportResult[];
-interface LintReportResult extends LintResult {
-  commitMessage: string;
-}
+import { LintReport, LintReportResult } from "./";
 
 const signs = {
   cross: "✖",
@@ -29,20 +24,16 @@ function formatResult(result: LintReportResult): string[] {
   return messageComponents;
 }
 
-export function formatResults(results: LintReportResults): string {
-  let violations = 0;
+export function formatReport(report: LintReport): string {
   let messageComponents: string[] = [];
-  for (const result of results) {
+  for (const result of report.results) {
     if (result.violations.length === 0) {
       continue;
     }
     const formattedResult = formatResult(result);
     messageComponents = messageComponents.concat(formattedResult, "");
-    violations += result.violations.length;
   }
-  if (violations > 0) {
-    const summaryText = chalk.bold(`${violations} problems found`);
-    messageComponents.push(`${chalk.red(signs.cross)}   ${summaryText}`);
-  }
+  const summaryText = chalk.bold(`${report.violationCount} problems found`);
+  messageComponents.push(`${chalk.red(signs.cross)}   ${summaryText}`);
   return messageComponents.join("\n");
 }
