+++ packages/cli/src/commands/lint/index.ts
@@ -1,51 +1,34 @@
-import fs from "fs";
-import path from "path";
-import getStdin from "get-stdin";
-import readCommits from "@commitlint/read";
 import { LintCommandFlags } from "../../types";
-import { lintCommitMessage } from "@oursky/gitlint";
-import { discoverConfig, applyPresets } from "@oursky/gitlint/lib/config";
+import { lintCommitMessage, LintResult } from "@oursky/gitlint";
 import { RulesPreset } from "@oursky/gitlint/lib/config/schema";
-import { formatResults } from "./format";
+import { loadPreset, loadCommitMessages } from "./loaders";
+import { formatReport } from "./format";
 
-const createConfigLoader = (configFile: string | undefined) => async (
-  _: string
-): Promise<string | null> => {
-  if (typeof configFile === "undefined") return null;
-  const filePath = path.resolve(".", configFile);
-  if (!fs.existsSync(filePath)) return null;
-  const stats = fs.statSync(filePath);
-  if (!stats.isFile()) return null;
-  return fs.readFileSync(filePath, "utf-8");
-};
-
-async function loadPreset(
-  configFile: string | undefined
-): Promise<RulesPreset> {
-  const loader = createConfigLoader(configFile);
-  const config = await discoverConfig(loader);
-  if (config === null && typeof configFile !== "undefined") {
-    throw new Error(`Config file not found: ${configFile}`);
-  }
-  return applyPresets(config);
+export interface LintReport {
+  violationCount: number;
+  results: LintReportResult[];
+}
+export interface LintReportResult extends LintResult {
+  commitMessage: string;
 }
 
-async function loadCommitMessages(flags: LintCommandFlags): Promise<string[]> {
-  if (typeof flags.stdIn !== "undefined" && !!flags.stdIn) {
-    const message = await getStdin();
-    if (message.length === 0) return [];
-    return [message];
-  } else if (
-    typeof flags.from !== "undefined" ||
-    typeof flags.to !== "undefined"
-  ) {
-    return readCommits({
-      from: flags.from,
-      to: flags.to,
+async function createLintReport(
+  messages: string[],
+  preset: RulesPreset
+): Promise<LintReport> {
+  const report: LintReport = {
+    results: [],
+    violationCount: 0,
+  };
+  for (const message of messages) {
+    const result = await lintCommitMessage(message, preset);
+    report.violationCount += result.violations.length;
+    report.results.push({
+      ...result,
+      commitMessage: message,
     });
   }
-  // Reads last commit from .git/COMMIT_EDITMSG
-  return readCommits({ edit: true });
+  return report;
 }
 
 async function lintCommand(flags: LintCommandFlags): Promise<void> {
@@ -55,17 +38,9 @@ async function lintCommand(flags: LintCommandFlags): Promise<void> {
     console.log("No commit messages found!");
     process.exit(0);
   }
-  const results = await Promise.all(
-    messages.map(async (message) => {
-      const result = await lintCommitMessage(message, preset);
-      return {
-        ...result,
-        commitMessage: message,
-      };
-    })
-  );
-  const formattedResults = formatResults(results);
-  if (formattedResults !== "") {
+  const report = await createLintReport(messages, preset);
+  if (report.violationCount > 0) {
+    const formattedResults = formatReport(report);
     console.log(formattedResults);
     process.exit(1);
   }
